ADD / Week6 / 표 병합(마무리).cpp, 표 병합 결과 사진
