+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "files.associations": {
+        "xstring": "cpp",
+        "sstream": "cpp",
+        "utility": "cpp",
+        "vector": "cpp",
+        "string": "cpp"
+    }
+}
